Towards enabling -Werror in skia on Linux
Review URL: http://codereview.appspot.com/5516044/



git-svn-id: http://skia.googlecode.com/svn/trunk@2983 2bbb7eff-a529-9590-31e7-b0007b416f81

+++ src/effects/SkTableColorFilter.cpp
@@ -171,7 +171,7 @@ SkTable_ColorFilter::SkTable_ColorFilter(SkFlattenableReadBuffer& buffer) : INHE
     size_t raw = SkPackBits::Unpack8(storage, size, fStorage);
 
     SkASSERT(raw <= sizeof(fStorage));
-    int count = gCountNibBits[fFlags & 0xF];
+    size_t count = gCountNibBits[fFlags & 0xF];
     SkASSERT(raw == count * 256);
 }
